Parse and persist storage bonus
